requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.31.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dotenv==0.9.9
fastapi==0.120.4
h11==0.16.0
idna==3.11
iso8601==2.1.0
isodate==0.7.2
kombu==5.5.4
lxml==6.0.2
packaging==25.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
pydantic==2.12.3
pydantic_core==2.41.4
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
zeep==4.3.2
